+++ Infrastructure/Repositories/ReportingRepository.cs
@@ -657,7 +657,7 @@ ORDER BY
                                 MaterialDescription = reader.IsDBNull(reader.GetOrdinal("MaterialDescription")) ? string.Empty : reader.GetString(reader.GetOrdinal("MaterialDescription")),
                                 Batch = reader.IsDBNull(reader.GetOrdinal("Batch")) ? string.Empty : reader.GetString(reader.GetOrdinal("Batch")),
                                 OrderQty = reader.IsDBNull(reader.GetOrdinal("OrderQty")) ? 0 : reader.GetInt32(reader.GetOrdinal("OrderQty")),
-                                CurQTY = reader.IsDBNull(reader.GetOrdinal("CurQTY")) ? 0 : Convert.ToDecimal(reader.GetValue(reader.GetOrdinal("CurQTY"))),
+                                CurQTY = reader.IsDBNull(reader.GetOrdinal("CurQTY")) ? 0 : Convert.ToInt32(reader.GetValue(reader.GetOrdinal("CurQTY"))),
                                 IssueCount = reader.IsDBNull(reader.GetOrdinal("IssueCount")) ? 0 : reader.GetInt32(reader.GetOrdinal("IssueCount")),
                                 ReceiptCount = reader.IsDBNull(reader.GetOrdinal("ReceiptCount")) ? 0 : reader.GetInt32(reader.GetOrdinal("ReceiptCount")),
                                 Deviation = reader.IsDBNull(reader.GetOrdinal("Deviation")) ? 0 : reader.GetInt32(reader.GetOrdinal("Deviation"))
